# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de Controle de Filmes - SENAI Jandira
  description: |-
    API destinada para controle de filmes
  termsOfService: http://swagger.io/terms/
  contact:
    email: 09matheushenry@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
paths:
...
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: Adiciona um novo Filme
      requestBody:
        description: JSON contendo os dados do novo filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
            example:
              nome: "Matheus contra os Malignos"
              data_lancamento: "2000-01-01"
              duracao: "01:40:00"
              sinopse: "teste 1"
              avaliacao: 5
              valor: 10
              capa: "malignos2"
              id_classificacao: 2
              genero:
                - id: 1
                - id: 2
              ator:
                - id: 1
                - id: 2
              diretor:
                - id: 1
                - id: 2
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Filme
      summary: Lista todos os Filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilmeGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um Filme pelo ID
      description: Retorna um Filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmeGet'
        '400':
          description: Error Bad Request (id)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Filme
      summary: Atualiza um filme pelo ID
      description: Atualiza um filme pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para atualizar o filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um filme existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
            example:
              nome: "Matheus contra os Malignos"
              data_lancamento: "2000-01-01"
              duracao: "01:40:00"
              sinopse: "teste 1"
              avaliacao: 5
              valor: 10
              capa: "malignos2"
              id_classificacao: 2
              genero:
                - id: 1
                - id: 2
              ator:
                - id: 1
                - id: 2
              diretor:
                - id: 1
                - id: 2
        required: true
      responses:
        '200':
          description: ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Filme
      summary: Exclui um Filme pelo ID
      description: Excluir um Filme existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── DIRETOR ───────────────────────────
...
  /v1/senai/locadora/sexo:
    post:
      tags:
        - Sexo
      summary: Adiciona um novo Sexo
      description: Adiciona um novo Sexo
      requestBody:
        description: JSON contendo os dados do Sexo a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
            example:
              sexo: "Feminino"
              sigla: "F"
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Sexo
      summary: Lista todos os Sexos
      description: Listagem de todos os Sexos registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SexoGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/sexo/{id}:
    get:
      tags:
        - Sexo
      summary: Busca um Sexo pelo ID
      description: Retorna um Sexo pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca do Sexo
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SexoGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Sexo
      summary: Atualiza um Sexo pelo ID
      description: Atualiza um Sexo pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para fazer a busca pelo sexo
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar um sexo existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
            example:
              sexo: "Feminino"
              sigla: "F"
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Sexo
      summary: Exclui um Sexo pelo ID
      description: Excluir um Sexo existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão do sexo
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── NACIONALIDADE ───────────────────────────
  /v1/senai/locadora/nacionalidade:
    post:
      tags:
        - Nacionalidade
      summary: Adiciona uma nova Nacionalidade
      description: Adiciona uma nova Nacionalidade
      requestBody:
        description: JSON contendo os dados da Nacionalidade a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
            example:
              nacionalidade: "Brasileiro"
              sigla: "BR"
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Nacionalidade
      summary: Lista todas as Nacionalidades
      description: Listagem de todas as Nacionalidades registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NacionalidadeGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/nacionalidade/{id}:
    get:
      tags:
        - Nacionalidade
      summary: Busca uma Nacionalidade pelo ID
      description: Retorna uma Nacionalidade pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da Nacionalidade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NacionalidadeGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Nacionalidade
      summary: Atualiza uma Nacionalidade pelo ID
      description: Atualiza uma Nacionalidade pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para fazer a busca pela nacionalidade
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma nacionalidade existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
            example:
              nacionalidade: "Brasileiro"
              sigla: "BR"
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Nacionalidade
      summary: Exclui uma Nacionalidade pelo ID
      description: Excluir uma Nacionalidade existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão da nacionalidade
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

  # ─────────────────────────── ATIVIDADE ───────────────────────────
...
  /v1/senai/locadora/foto:
    post:
      tags:
        - Foto
      summary: Adiciona uma nova Foto
      description: Adiciona uma nova Foto
      requestBody:
        description: JSON contendo os dados da Foto a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Foto'
            example:
              foto_url: "https://exemplo.com/fotos/camila.jpg"
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Foto
      summary: Lista todas as Fotos
      description: Listagem de todas as Fotos registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FotoGet'
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
  /v1/senai/locadora/foto/{id}:
    get:
      tags:
        - Foto
      summary: Busca uma Foto pelo ID
      description: Retorna uma Foto pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca da Foto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FotoGet'
        '400':
          description: Error Bad Request (ID)
        '404':
          description: Error Not found
        '500':
          description: Error Internal Server (Controller | Model)
    put:
      tags:
        - Foto
      summary: Atualiza uma Foto pelo ID
      description: Atualiza uma Foto pelo ID colocando as informações da mesma forma que o POST
      parameters:
        - name: id
          in: path
          description: ID para fazer a busca pela foto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar uma foto existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Foto'
            example:
              foto_url: "https://exemplo.com/fotos/camila.jpg"
        required: true
      responses:
        '200':
          description: Ok
        '400':
          description: Error Bad Request (id | dados)
        '404':
          description: Error Not Found
        '415':
          description: Error Content Type
        '500':
          description: Error internal server (Controller | Model)
    delete:
      tags:
        - Foto
      summary: Exclui uma Foto pelo ID
      description: Excluir uma Foto existente buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para realizar a busca para a exclusão da foto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success Delete
        '400':
          description: Error Bad Request (ID)
        '500':
          description: Error Internal Server (Controller | Model)

# ═══════════════════════════════════════════════════════════════
#  SCHEMAS
# ═══════════════════════════════════════════════════════════════
components:
  schemas:
...
    IdRef:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1

    # ── FILME ──
    Filme:
      required:
        - nome
        - sinopse
        - capa
        - data_lancamento
        - duracao
        - id_classificacao
      type: object
      properties:
        nome:
          type: string
          maxLength: 80
          example: "Matheus contra os Malignos"
          description: Máximo de 80 caracteres
        sinopse:
          type: string
          example: "teste 1"
          description: Sem limite de caracteres
        capa:
          type: string
          maxLength: 255
          example: "malignos2"
          description: Máximo de 255 caracteres
        data_lancamento:
          type: string
          format: date
          example: "2000-01-01"
        duracao:
          type: string
          example: "01:40:00"
          pattern: '^(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d$'
          description: Formato HH:MM:SS
        valor:
          type: number
          format: double
          example: 10.00
          description: Máximo 999.99 (duas casas decimais)
        avaliacao:
          type: number
          format: double
          example: 5.0
          description: De 0.0 a 5.0 (uma casa decimal)
        id_classificacao:
          type: integer
          format: int64
          example: 2
          description: ID da classificação indicativa
        genero:
          type: array
          description: Lista de IDs de gêneros associados ao filme
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 1
            - id: 2
        ator:
          type: array
          description: Lista de IDs de atores associados ao filme
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 1
            - id: 2
        diretor:
          type: array
          description: Lista de IDs de diretores associados ao filme
          items:
            $ref: '#/components/schemas/IdRef'
          example:
            - id: 1
            - id: 2
    FilmeGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          maxLength: 80
          example: "Matheus contra os Malignos"
        sinopse:
          type: string
          example: "teste 1"
        capa:
          type: string
          maxLength: 255
          example: "malignos2"
        data_lancamento:
          type: string
          format: date
          example: "2000-01-01"
        duracao:
          type: string
          example: "01:40:00"
        valor:
          type: number
          format: double
          example: 10.00
        avaliacao:
          type: number
          format: double
          example: 5.0
        id_classificacao:
          type: integer
          format: int64
          example: 2
        genero:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'
        ator:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'
        diretor:
          type: array
          items:
            $ref: '#/components/schemas/IdRef'

    # ── DIRETOR ──
...
    Sexo:
      required:
        - sexo
        - sigla
      type: object
      properties:
        sexo:
          type: string
          maxLength: 20
          example: "Feminino"
          description: Máximo de 20 caracteres
        sigla:
          type: string
          maxLength: 3
          example: "F"
          description: Máximo de 3 caracteres
    SexoGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        sexo:
          type: string
          maxLength: 20
          example: "Masculino"
        sigla:
          type: string
          maxLength: 3
          example: "M"

    # ── NACIONALIDADE ──
    Nacionalidade:
      required:
        - nacionalidade
        - sigla
      type: object
      properties:
        nacionalidade:
          type: string
          maxLength: 90
          example: "Brasileiro"
          description: Máximo de 90 caracteres
        sigla:
          type: string
          maxLength: 4
          example: "BR"
          description: Máximo de 4 caracteres
    NacionalidadeGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nacionalidade:
          type: string
          maxLength: 90
          example: "Brasileiro"
        sigla:
          type: string
          maxLength: 4
          example: "BR"

    # ── ATIVIDADE ──
...
    Foto:
      required:
        - foto_url
      type: object
      properties:
        foto_url:
          type: string
          maxLength: 255
          example: "https://exemplo.com/fotos/camila.jpg"
          description: Máximo de 255 caracteres
    FotoGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        foto_url:
          type: string
          maxLength: 255
          example: "https://exemplo.com/fotos/camila.jpg"

  # ── Request Bodies reutilizáveis ──
...